+++ MaterialDesignExtensions/Model/FirstLevelNavigationItem.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MaterialDesignExtensions.Model
+{
+    public class FirstLevelNavigationItem : NavigationItem
+    {
+        public FirstLevelNavigationItem() : base() { }
+    }
+}